requirements.txt
numpy
tifffile
pynrrd
pillow